+++ ois_api_client/dto/DiscountData.py
@@ -1,6 +1,3 @@
-from decimal import Decimal
-
-
 class DiscountData:
     """Discount data
     :param discount_description: Description of the discount
@@ -8,7 +5,7 @@ class DiscountData:
     :param discount_rate: Rate of discount per item expressed in the currency of the invoice if not included in the unit price
     """
 
-    def __init__(self, discount_description: str, discount_value: Decimal, discount_rate: Decimal):
+    def __init__(self, discount_description: str, discount_value: float, discount_rate: float):
         self.discount_description = discount_description
         self.discount_value = discount_value
         self.discount_rate = discount_rate
