+++ ois_api_client/dto/LineAmountsSimplified.py
@@ -1,6 +1,3 @@
-from decimal import Decimal
-
-
 class LineAmountsSimplified:
     """Item value data to be completed in case of simplified invoice
 
@@ -9,8 +6,8 @@ class LineAmountsSimplified:
     :param line_gross_amount_simplified_huf:
     """
 
-    def __init__(self, line_vat_content: Decimal, line_gross_amount_simplified: Decimal,
-                 line_gross_amount_simplified_huf: Decimal):
+    def __init__(self, line_vat_content: float, line_gross_amount_simplified: float,
+                 line_gross_amount_simplified_huf: float):
         self.line_vat_content = line_vat_content
         self.line_gross_amount_simplified = line_gross_amount_simplified
         self.line_gross_amount_simplified_huf = line_gross_amount_simplified_huf
