+++ ois_api_client/dto/ProductFeeData.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from .ProductCode import ProductCode
 from .ProductFeeMeasuringUnit import ProductFeeMeasuringUnit
 
@@ -14,9 +12,9 @@ class ProductFeeData:
     :param product_fee_amount: Amount in Hungarian forints of the product fee
     """
 
-    def __init__(self, product_fee_code: ProductCode, product_fee_quantity: Decimal,
-                 product_fee_measuring_unit: ProductFeeMeasuringUnit, product_fee_rate: Decimal,
-                 product_fee_amount: Decimal):
+    def __init__(self, product_fee_code: ProductCode, product_fee_quantity: float,
+                 product_fee_measuring_unit: ProductFeeMeasuringUnit, product_fee_rate: float,
+                 product_fee_amount: float):
         self.product_fee_code = product_fee_code
         self.product_fee_quantity = product_fee_quantity
         self.product_fee_measuring_unit = product_fee_measuring_unit
