+++ ois_api_client/dto/ProductFeeTakeoverData.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from .Takeover import Takeover
 
 
@@ -10,6 +8,6 @@ class ProductFeeTakeoverData:
     :param takeover_amount: Amount in Hungarian forints of the product fee taken over if the purchaser takes over the supplier’s product fee liability
     """
 
-    def __init__(self, takeover_reason: Takeover, takeover_amount: Decimal):
+    def __init__(self, takeover_reason: Takeover, takeover_amount: float):
         self.takeover_reason = takeover_reason
         self.takeover_amount = takeover_amount
