+++ ois_api_client/dto/SummaryGrossData.py
@@ -1,6 +1,3 @@
-from decimal import Decimal
-
-
 class SummaryGrossData:
     """Gross data of the invoice summary
 
@@ -8,6 +5,6 @@ class SummaryGrossData:
     :param invoice_gross_amount_huf: Gross amount of the invoice expressed in HUF
     """
 
-    def __init__(self, invoice_gross_amount: Decimal, invoice_gross_amount_huf: Decimal):
+    def __init__(self, invoice_gross_amount: float, invoice_gross_amount_huf: float):
         self.invoice_gross_amount = invoice_gross_amount
         self.invoice_gross_amount_huf = invoice_gross_amount_huf
