+++ ois_api_client/dto/SummarySimplified.py
@@ -1,6 +1,3 @@
-from decimal import Decimal
-
-
 class SummarySimplified:
     """Calculation of simplified invoice totals
     :param vat_content: In the case of a simplified invoice, VAT content rate
@@ -8,7 +5,7 @@ class SummarySimplified:
     :param vat_content_gross_amount_huf: The gross amount of the sale or service for the given tax amount in HUF
     """
 
-    def __init__(self, vat_content: Decimal, vat_content_gross_amount: Decimal, vat_content_gross_amount_huf: Decimal):
+    def __init__(self, vat_content: float, vat_content_gross_amount: float, vat_content_gross_amount_huf: float):
         self.vat_content = vat_content
         self.vat_content_gross_amount = vat_content_gross_amount
         self.vat_content_gross_amount_huf = vat_content_gross_amount_huf
