+++ ois_api_client/dto/VatRateGrossData.py
@@ -1,6 +1,3 @@
-from decimal import Decimal
-
-
 class VatRateGrossData:
     """Gross data of given tax rate
 
@@ -8,6 +5,6 @@ class VatRateGrossData:
     :param vat_rate_gross_amount_huf: Gross amount of sales or service delivery under a given tax rate expressed in HUF
     """
 
-    def __init__(self, vat_rate_gross_amount: Decimal, vat_rate_gross_amount_huf: Decimal):
+    def __init__(self, vat_rate_gross_amount: float, vat_rate_gross_amount_huf: float):
         self.vat_rate_gross_amount = vat_rate_gross_amount
         self.vat_rate_gross_amount_huf = vat_rate_gross_amount_huf
