+++ ois_api_client/dto/VatRateNetData.py
@@ -1,6 +1,3 @@
-from decimal import Decimal
-
-
 class VatRateNetData:
     """Net data of given tax rate
 
@@ -8,6 +5,6 @@ class VatRateNetData:
     :param vat_rate_net_amount_huf: Net amount of sales or service delivery under a given tax rate expressed in HUF
     """
 
-    def __init__(self, vat_rate_net_amount: Decimal, vat_rate_net_amount_huf: Decimal):
+    def __init__(self, vat_rate_net_amount: float, vat_rate_net_amount_huf: float):
         self.vat_rate_net_amount = vat_rate_net_amount
         self.vat_rate_net_amount_huf = vat_rate_net_amount_huf
